+++ feature_modules/feature_module_base/lib/src/feature_module_base_module.dart
@@ -1,1 +1,2 @@
 export 'package:app/app.dart';
+export 'package:splash/splash.dart';